Implement a 8pt grid system
Using a 8pt grid system will add a consistent rule to how elements are
sized / laid out. It creates a natural flow for the content which is
aesthetically pleasing. Read more here:
https://builttoadapt.io/intro-to-the-8-point-grid-system-d2573cde8632

Also reorder the CSS properties to be in alphabetical order. This will
consistent rule will ensure it is easier to maintain the stylesheet.

Remove redundant CSS rules that is providing no value to the markup.